[okera] add injectable authz implementation
This change adds just enough code to allow injecting a policy engine into
Minio from a different package. Specifically, it defines the PolicyEngine
interface (and a mechanism to set the global variable, following the existing
Minio code style).

This interface implements the functions necessary (namely getting the
credentials for a specific user and whether a particular operation
is allowed) for the relevant parts of the IAM subsystem to light up.

In addition to the interface, a little bit of plumbing is added to
add calls to the injected policy engine, as well as a wrapper implementation
of IAMStorageAPI since the methods on this interface are all private
and can't be implemented outside the package.

The goal is that the Minio side of the code won't be needed to be changed
frequently and the core of the implementation is in the Okera codebase.